+++ src/Forge.Forms.Mapper/Forge.Forms.Mapper.csproj
@@ -38,18 +38,20 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Interceptors\MapperActionInterceptor.cs" />
+    <Compile Include="Interceptors\MapperInterceptor.cs" />
     <Compile Include="MaterialMapper.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="ModuleInitializer.cs" />
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="..\src\Forge.Forms\Forge.Forms.csproj">
+    <PackageReference Include="Proxier" Version="1.4.7" />
+  </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\Forge.Forms\Forge.Forms.csproj">
       <Project>{4e0c323d-5afc-4105-815f-587346aa5321}</Project>
       <Name>Forge.Forms</Name>
     </ProjectReference>
-  </ItemGroup>
-  <ItemGroup>
-    <PackageReference Include="Proxier" Version="1.4.7" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it.
