+++ src/Forge.Forms.Mapper/Interceptors/MapperInterceptor.cs
@@ -1,0 +1,14 @@
+﻿using Proxier.Mappers;
+
+namespace Forge.Forms.Mapper.Interceptors
+{
+    class MapperInterceptor : IModelInterceptor
+    {
+        public IModelContext Intercept(IModelContext modelContext)
+        {
+            return modelContext.NewModel == null
+                ? modelContext
+                : new ModelContext(modelContext.NewModel.GetInjectedObject(), modelContext.ResourceContext);
+        }
+    }
+}
